Add confirm modal before deleting photos in photo preview

+++ src/components/Camera/PhotoPreview.js
@@ -1,10 +1,11 @@
 // @flow
 
-import React from "react";
+import React, { useState } from "react";
 import { FlatList, Image, Pressable, Text } from "react-native";
 import type { Node } from "react";
 import RNFS from "react-native-fs";
-import { Avatar } from "react-native-paper";
+import { Button, Paragraph, Dialog, Portal, Avatar } from "react-native-paper";
+import { useTranslation } from "react-i18next";
 
 import { viewStyles, imageStyles, textStyles } from "../../styles/camera/standardCamera";
 import Photo from "../../models/Photo";
@@ -15,28 +16,59 @@ type Props = {
 }
 
 const PhotoPreview = ( { photos, setPhotos }: Props ): Node => {
+  const { t } = useTranslation( );
+  const [visible, setVisible] = useState( false );
+  const [photoToDelete, setPhotoToDelete] = useState( null );
+
+  const showDialog = ( ) => setVisible( true );
+  const hideDialog = ( ) => {
+    setPhotoToDelete( null );
+    setVisible( false );
+  };
+
+  const deletePhoto = ( ) => {
+    if ( !photoToDelete ) { return; }
+    const updatedPhotos = photos;
+    const photoIndex = photos.findIndex( p => p === photoToDelete );
+    updatedPhotos.splice( photoIndex, 1 );
+
+    // spreading the array forces PhotoPreview to rerender on each photo deletion
+    setPhotos( [...updatedPhotos] );
+
+    // delete photo thumbnail from temp directory
+    const tempDirectory = `${RNFS.TemporaryDirectoryPath}/ReactNative`;
+    const fileName = photoToDelete.path.split( "ReactNative/" )[1];
+    RNFS.unlink( `${tempDirectory}/${fileName}` );
+
+    hideDialog( );
+  };
+
   const renderSmallPhoto = ( { item } ) => {
     const uri = Photo.setPlatformSpecificFilePath( item.path );
 
-    const deletePhoto = ( ) => {
-      const updatedPhotos = photos;
-      const photoIndex = photos.findIndex( p => p === item );
-      updatedPhotos.splice( photoIndex, 1 );
-
-      // spreading the array forces PhotoPreview to rerender on each photo deletion
-      setPhotos( [...updatedPhotos] );
-
-      // delete photo thumbnail from temp directory
-      const tempDirectory = `${RNFS.TemporaryDirectoryPath}/ReactNative`;
-      const fileName = item.path.split( "ReactNative/" )[1];
-      RNFS.unlink( `${tempDirectory}/${fileName}` );
-    };
-
     return (
       <>
+        <Portal>
+          <Dialog visible={visible} onDismiss={hideDialog}>
+            <Dialog.Content>
+              <Paragraph>{t( "Are-you-sure" )}</Paragraph>
+            </Dialog.Content>
+            <Dialog.Actions>
+              <Button onPress={hideDialog} style={viewStyles.cancelButton}>
+                {t( "Cancel" )}
+              </Button>
+              <Button onPress={deletePhoto} style={viewStyles.confirmButton}>
+                {t( "Yes-delete-photo" )}
+              </Button>
+            </Dialog.Actions>
+          </Dialog>
+        </Portal>
         <Image source={{ uri }} style={imageStyles.smallPhoto} />
         <Pressable
-          onPress={deletePhoto}
+          onPress={( ) => {
+            setPhotoToDelete( item );
+            showDialog( );
+          }}
           style={viewStyles.deleteButton}
         >
           <Avatar.Icon icon="delete-forever" size={30} />
@@ -45,7 +77,11 @@ const PhotoPreview = ( { photos, setPhotos }: Props ): Node => {
     );
   };
 
-  const emptyDescription = ( ) => <Text style={textStyles.topPhotoText}>Photos you take will appear here</Text>;
+  const emptyDescription = ( ) => (
+    <Text style={textStyles.topPhotoText}>
+      {t( "Photos-you-take-will-appear-here" )}
+    </Text>
+  );
 
   return (
     <FlatList
